+++ src/PKSim.Core/Commands/ResetExpressionParameterCommand.cs
@@ -1,0 +1,23 @@
+using OSPSuite.Core.Domain;
+using PKSim.Core.Model;
+using PKSim.Core.Services;
+
+namespace PKSim.Core.Commands;
+
+public class ResetExpressionParameterCommand : ResetParameterCommand
+{
+   public ResetExpressionParameterCommand(IParameter parameter) : base(parameter)
+   {
+   }
+
+   protected override void UpdateDependenciesOnParameter(IParameter parameter, IExecutionContext context)
+   {
+      base.UpdateDependenciesOnParameter(parameter, context);
+
+      if (context.BuildingBlockContaining(parameter) is not ExpressionProfile expressionProfile)
+         return;
+
+      var updateTask = context.Resolve<IExpressionProfileUpdater>();
+      updateTask.SynchronizeAllSimulationSubjectsWithExpressionProfile(expressionProfile);
+   }
+}
